Fixing issues with inconsistent coordinate systems between the robot and dice

+++ Dice/src/dice/strategy/action/attacker/ToZoneCenterAction.java
@@ -41,7 +41,7 @@ public class ToZoneCenterAction extends StrategyAction {
 		}
 		
 		long angle = (long) Math.toDegrees(getTargetObject(state).getRotationRelativeTo(zoneCenter));
-		byte dist = (byte) getTargetObject(state).getEuclidean(zoneCenter);
+		double dist = getTargetObject(state).getEuclidean(zoneCenter);
 		
 		return RobotInstruction.CreateMoveTo(angle, dist);
 	}
